delets garbage

+++ src/main/java/com/spring/bioMedical/repository/AdminRepository.java
@@ -1,18 +1,16 @@
 package com.spring.bioMedical.repository;
 
-import java.util.List;
-
+import com.spring.bioMedical.entity.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.spring.bioMedical.entity.Admin;
-import com.spring.bioMedical.entity.User;
+import java.util.List;
 
 
 @Repository
-public interface AdminRepository extends JpaRepository<Admin, Long> {
+public interface AdminRepository extends JpaRepository<Admin, Integer> {
 
-	Admin findByEmail(String user);
-	
-	List<Admin> findByRole(String user);
+    Admin findByEmail(String user);
+
+    List<Admin> findByRole(String user);
 }
